refactor: revert mmsi columns to INTEGER and restore README
Change the mmsi column type back from BIGINT to INTEGER in the
create-table and insert SQL for performance, and align the CSV reader
and db.rs typing accordingly. Replace pypi_readme.rst with a restored
README referenced from pyproject.toml, tidy decoder.py and the worker
logic in src/lib.rs, and drop the all-contributors config and
CONTRIBUTORS.md.

+++ aisdb_lib/src/db.rs
@@ -266,7 +266,7 @@ pub fn postgres_insert_dynamic(
         let _ = tx.execute(
             &stmt,
             &[
-                &p.mmsi,
+                &(p.mmsi as i32),
                 &(e as i32),
                 &(p.longitude.unwrap_or_default() as f32),
                 &(p.latitude.unwrap_or_default() as f32),
